Fixed fault in Sub60StartNewWarpCommand
Had incorrect unused size. Byte -> short on 4th member.

+++ src/Booma.Proxy.Packets.BlockServer/Commands/Command60/Sub60StartNewWarpCommand.cs
@@ -30,7 +30,7 @@ namespace Booma.Proxy
 		
 		//TODO: What is this?
 		[WireMember(4)]
-		public byte Unused2 { get; }
+		public short Unused2 { get; }
 
 		public Sub60StartNewWarpCommand()
 		{
